extract program setup to config methods
